change response codes to 500 for splash communication errors
Technically, for the user the communication with the splash container is
an implementation detail of the server. Returning a 502 (Bad Gateway)
should indicate to the user, that the resource (in this case the url
that was transmitted to the extract endpoint) is no longer available,
whereas internal communication problems should not give this indication
to the user.