introduce Content class which delays splash IO to extraction phase
The Content class essentially replaces the WebsiteData class. It can
be initialized with a URL only and provides async getters for various
attributes that are currently used by multiple extractors. These getters
are lazy, i.e. they will only perform work once actually invoked and
then store the result in the Content for usage by other extractors.
This allows to move the communication with Splash from the
MetadataManager into one of those lazy getters, and hence into the
extration phase. This turns the previously sequential communication
pattern with splash and lighthouse into a concurrent one and saves up to
10 seconds in total response time.