replace outdated adblockparser dependency with newer adblock
The new adblock dependency is a wrapper around a rust library. It is
expected to perform much better than the regex based old package and
using it removes a couple of workarounds that were meant to speed up the
old implementation (e.g. the google-re2 hack).

In the course of exchanging the dependency, this commit also changes the
behaviour of the actual rule based extractors:
 - They no longer use the "raw-links" extracted from the received html.
 - Instead they use the requests recorded when loading the page.
This means, that a lot less urls need to be checked against the adblock
rules, which should further speed up the analysis.
This also means, that a static link that is present in the html, that
previously would have triggered the extractor will no longer do so.